#36 Refactor SSH handlers
